1) Refactored Project to Feature-sliced design , 2) created AppRouter config
