+++ src/app/providers/ThemeProvider/index.tsx
@@ -1,0 +1,8 @@
+import ThemeProvider from "./ui/ThemeProvider";
+import { useTheme } from "./lib/useTheme";
+
+
+export {
+    ThemeProvider,
+    useTheme
+}
