+++ src/app/providers/router/index.tsx
@@ -1,0 +1,1 @@
+export { AppRouter } from './ui/AppRouter';
