+++ src/app/providers/router/ui/AppRouter.tsx
@@ -1,0 +1,21 @@
+import React, {Suspense} from 'react';
+import {Route, Routes} from "react-router-dom";
+import {MainPageAsync} from "pages/MainPage";
+import {AboutPageAsync} from "pages/AboutPage";
+import {routeConfig} from "shared/config/ routeConfig/routeConfig";
+
+export const AppRouter = () => {
+    return (
+        <Suspense fallback={<div>Loading...</div>}>
+            <Routes>
+                {Object.values(routeConfig).map(({element, path}) => (
+                    <Route
+                        key={path}
+                        path={path}
+                        element={element}
+                    />
+                    ))}
+            </Routes>
+        </Suspense>
+    );
+};
