--- src/components/Counter.module.css
@@ -1,6 +1,0 @@
-button {
-  padding: 20px;
-  color: aqua;
-}
-
-/*# sourceMappingURL=Counter.module.css.map */
