+++ src/pages/AboutPage/index.ts
@@ -1,0 +1,5 @@
+export { AboutPageAsync } from "./ui/AboutPage.async";
+//
+// export {
+//     AboutPageAsync as AboutPage
+// }
