+++ src/pages/MainPage/index.tsx
@@ -1,0 +1,1 @@
+export { MainPageAsync } from './ui/MainPage.async'
