+++ src/pages/MainPage/ui/MainPage.async.tsx
