+++ src/shared/config/ routeConfig/routeConfig.tsx
@@ -1,0 +1,26 @@
+import {RouteProps} from "react-router-dom";
+import MainPage from "pages/MainPage/ui/MainPage";
+import AboutPage from "pages/AboutPage/ui/AboutPage";
+
+export enum AppRoutes {
+    MAIN = 'main',
+    ABOUT = 'about',
+
+}
+
+export const RoutePath: Record<AppRoutes, string> = {
+    [AppRoutes.MAIN]: '/',
+    [AppRoutes.ABOUT]: '/about',
+}
+
+
+export const routeConfig: Record<AppRoutes, RouteProps> = {
+    [AppRoutes.MAIN]: {
+        path: RoutePath.main,
+        element: <MainPage />
+    },
+    [AppRoutes.ABOUT]: {
+        path: RoutePath.about,
+        element: <AboutPage />
+    }
+}
